feat(wip): moved link parser to common module
