+++ common/src/main/java/edu/link/parsers/GitHubLinkParser.java
@@ -1,19 +1,19 @@
-package edu.java.bot.link.parsers;
+package edu.link.parsers;
 
-import edu.java.bot.link.links.Link;
-import edu.java.bot.link.links.StackOverflowLink;
+import edu.link.links.GitHubLink;
+import edu.link.links.Link;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-public class StackOverflowLinkParser extends LinkParser {
+public class GitHubLinkParser extends LinkParser {
     private final static Pattern URL_PATTERN =
-        Pattern.compile("https://stackoverflow.com/questions/(?<questionId>\\d*)/\\S*");
+        Pattern.compile("https://github\\.com/(?<userName>\\S*)/(?<repoName>\\S*)");
 
     public Link parse(String url) {
         Matcher matcher = URL_PATTERN.matcher(url);
 
         if (matcher.find()) {
-            return new StackOverflowLink(url, Long.parseLong(matcher.group("questionId")));
+            return new GitHubLink(url, matcher.group("userName"), matcher.group("repoName"));
         } else if (next != null) {
             return next.parse(url);
         } else {
