+++ common/src/main/java/edu/link/parsers/StackOverflowLinkParser.java
@@ -1,19 +1,20 @@
-package edu.java.bot.link.parsers;
+package edu.link.parsers;
 
-import edu.java.bot.link.links.GitHubLink;
-import edu.java.bot.link.links.Link;
+import edu.link.links.Link;
+import edu.link.links.StackOverflowLink;
+
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-public class GitHubLinkParser extends LinkParser {
+public class StackOverflowLinkParser extends LinkParser {
     private final static Pattern URL_PATTERN =
-        Pattern.compile("https://github\\.com/(?<userName>\\S*)/(?<repoName>\\S*)");
+        Pattern.compile("https://stackoverflow.com/questions/(?<questionId>\\d*)/\\S*");
 
     public Link parse(String url) {
         Matcher matcher = URL_PATTERN.matcher(url);
 
         if (matcher.find()) {
-            return new GitHubLink(url, matcher.group("userName"), matcher.group("repoName"));
+            return new StackOverflowLink(url, Long.parseLong(matcher.group("questionId")));
         } else if (next != null) {
             return next.parse(url);
         } else {
